Fix deadlock
The pg-query-stream won't let other queries occur until all the data
has been read. That's fine for e.g. publishing but we can't do anything
while processing a message (because then we'd block the query stream
and that's the deadlock). Moreover, since stuff might be on our queue,
we can't push or unshift to it.
The only thing we do when processing a message is for single messages -
and we have a lock client for that anyway. This change removes doing that
in our queue and uses the lock client for checking if the message still
exists.